requirements.txt
django==3.0.4
psycopg2==2.8.4
Pillow==7.0.0
gunicorn==20.0.4
google-api-python-client==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1